♻️refactor: model => controller로 반영
