+++ src/app/log/record/components/LogRecordEdit/LogRecordEdit.controller.ts
@@ -1,18 +1,30 @@
-import { useEffect, useMemo } from "react";
-import LogRecordEditView from "./LogRecordEdit.view";
+"use client";
+
+import { useEffect } from "react";
 import useUserLocationStore from "@/stores/useUserLocationStore";
 import { useUI } from "@/components/uiContext/UiContext";
 import useLogRecordContext from "../../context/LogRecordContext";
+import { useMutation } from "@tanstack/react-query";
+import { MASIL_KEY } from "@/lib/api/queryKeys";
+import { postMasil } from "@/lib/api/masil/server";
+import { MasilRecordRequest } from "@/types/Request";
+import { drawPath } from "@/utils/drawPath";
+import { useRouter } from "next/navigation";
 import { useForm } from "react-hook-form";
 
 const useLogRecordEditController = () => {
   const { setModalView, openModal, closeModal } = useUI();
   const { setUserLocation } = useUserLocationStore();
-
   const { logData, handleRemovePin, handleClickPin } = useLogRecordContext();
+  const { register, getValues } = useForm<{ logMemo: string }>({
+    defaultValues: { logMemo: logData.content },
+  });
+  const router = useRouter();
 
-  const { register, watch } = useForm();
-  const watchLogMemo = useMemo(() => watch("logMemo"), [watch("logMemo")]);
+  const mutation = useMutation({
+    mutationKey: [MASIL_KEY.RECORD_SUBMIT],
+    mutationFn: async (data: MasilRecordRequest) => postMasil({ data }),
+  });
 
   useEffect(() => {
     const pathLength = logData.path.length;
@@ -26,61 +38,76 @@ const useLogRecordEditController = () => {
   }, []);
 
   /**
-   * @func handleImageUpload
-   * @params (image: File, pinIndex: number)
-   * @brief 사용자가 핀 썸네일을 첨부하면 실행됩니다. API와 통신하여 이미지 파일을 서버에 저장, URL로 변환하여 반환받습니다. 성공적으로 반환받은 경우 썸네일 URL을 logData의 핀 인덱스에 저장합니다.
-   */
-  const handleImageUpload = (pinIndex: number, image: File) => {
-    // TODO: ImageFile 전송 후 URL 반환 (common 분리?)
-    // PinEditModal에서 구현할수도 있음
-  };
-
-  /**
    * @func handleSubmit
-   * @params (memo: string)
+   *
    * @brief 폼을 통해 입력받은 Memo를 logData의 content에 저장한 후, 서버에 전송합니다.
+   *
+   * 이후, 서버에 전송 성공시 Done Modal이 제공되어지고, 공유를 선택시 시작지점과 성성된 log id 를 post 생성 페이지에 전달합니다.
    */
-  const handleSubmit = (memo: string) => {
+  const handleSubmit = () => {
     // const pathThumbnailUrl = drawPath(logData.path);
-
-    // 경로를 기반으로 그려진 썸네일과 메모를 logData 데이터에 저장
-    // setLogData((prevData) => {
-    //   return { ...prevData, thumbnailUrl: pathThumbnailUrl, content: memo };
-    // });
 
     // TODO: logData API 통신
 
-    // API 통신 완료 시
-    setModalView("LOG_RECORD_DONE_VIEW");
-    openModal({
-      onClickUploadPost: () => {
-        // 산책로 공유하기를 클릭한 경우
-        // 산책로 공유하기 페이지로 이동, logData를 전달
-        closeModal();
+    const test = {
+      depth1: "세종",
+      depth2: "대평",
+      depth3: "24",
+      depth4: "12421",
+      path: [
+        { lat: 22, lng: 22 },
+        { lat: 22, lng: 22 },
+        { lat: 22, lng: 22 },
+        { lat: 22, lng: 22 },
+      ],
+      title: "4214",
+      content: getValues("logMemo"),
+      distance: 0,
+      totalTime: 0,
+      startedAt: new Date().toISOString(),
+      pins: [],
+      thumbnailUrl: "ㅇㅇㅇ",
+      postId: null,
+    };
+
+    mutation.mutate(test, {
+      onSuccess: (res) => {
+        if (!res) {
+          return;
+        }
+        const { id } = res;
+        router.push(`/log/${id}`);
+
+        setModalView("LOG_RECORD_DONE_VIEW");
+        openModal({
+          onClickUploadPost: () => {
+            const startPoint = JSON.stringify(logData.path[0]);
+
+            router.push(`/post/create?id=${id}&point=${startPoint}`);
+            closeModal();
+          },
+          onClickCancel: () => {
+            closeModal();
+          },
+          logData,
+        });
       },
-      onClickCancle: () => {
-        // 산책로 공유하기를 클릭하지 않은 경우
-        // 메인 페이지 혹은 내 기록 페이지로 이동
-        closeModal();
+
+      onError: () => {
+        setModalView("LOG_RECORD_ALERT_VIEW");
+        openModal({
+          message: `산책로 저장에 오류가 발생하였습니다. 잠시 후 다시 시도해주세요.`,
+        });
       },
-      logData,
     });
-
-    // setLogData(DEFAULT_LOG_DATA);
-
-    // navigator.geolocation.getCurrentPosition(({ coords }) => {
-    //   setUserLocation({ lat: coords.latitude, lng: coords.longitude });
-    // });
   };
 
   return {
     logData,
-    watchLogMemo,
 
     register,
     handleClickPin,
     handleRemovePin,
-    handleImageUpload,
     handleSubmit,
   };
 };
