+++ src/app/log/record/components/index.ts
@@ -1,3 +1,3 @@
 export { default as LogRecordStandby } from "./LogRecordStandby/LogRecordStandby.view";
 export { default as LogRecordRecording } from "./LogRecordRecording/LogRecordRecording.view";
-export { default as LogRecordEdit } from "./LogRecordEdit/LogRecordEdit.model";
+export { default as LogRecordEdit } from "./LogRecordEdit/LogRecordEdit.view";
